RELEASE 1.2.0
Increased minimum fractional improvement for reporting a new best
individual to 2%. Reduces the log chatter and speeds things up when
reporting (e.g., plot generation) constitutes a significant part of
the overall CPU load.

Finally got abort working (with Enter key)! Possibly some of the
checks for self.running in DifferentialEvolution may be superfluous
now, but they don't hurt and can stay in. Has Python ignore the ^C
signal since it doesn't work properly anyhow.

Dependency on AsynQueue >= 0.9.7, which traps an exception raised by a
multiprocessing pipe being broken. That can happen when a ProcessQueue
is used, as will often be the case, and it gets shut down while
evaluations are still pending. The latest AsynQueue release, in
conjunction with this one for ade, shuts down early very nicely.