RELEASE 1.2.2
Since 1.2.1:
---------------------------------------------------------------------
Smoother concurrency-limited population setup, with no orphaned
evaluations.

Enter-to-abort now works consistently and responsively, during
Population setup and during the DifferentialEvolution run.

Improvements to the lgg paramfinder logfile parsing script.

Any object that can have float(...) called on it (or a Deferred that
fires with such an object) may be returned from the evaluation
function as the SSE.