RELEASE 1.3.1
Since 1.3.0
-----------------------------------------------------------------------
Major updates to Population history analysis:

(1) Saves, loads, and analyzes failed challenges when not using xSSE
option. In analysis, updated color scheme to clearly show and
differentiate now-in-population, once-in-population, and
never-in-population.  SSE+param combos never in a population are
plotted with small grey dots.

(2) Selective retention of SSE+param combos under history size
limiting favors keeping combos with lower SSE, and that have been in a
population. Vastly improved weighting of once-in-population vs
never-in-population to keep once-in-population well represented in
history.

(3) XY cross-correlation plots include a best-fit line with slope and
y-intercept, essential for specifying relation constraints in a
related project.

All of this (1-3) is accessible with a convenient command-line
tool, 'pv' (for "parameter visualization."

Docstring updates.

Call to DifferentialEvolution instance can now include a callback
'func' (keyword) that gets called with generation number (int) after
every generation.